Batch update from work PC:
- Logo for each window of the application.
- TL currency formatting for money related UI controls.
- WindowStartupLocation code is moved from the parent code-behind to the child xaml.
- Deal TotalPrice is changed with UnitPrice (was a bug which caused a mix of unit & total price in the product details window).
- Removing a specific expiration date of a product is possible now.
- Column sorting bugfix for DataGridTemplateColumns.
- Many improvements in purchase list window.
- Bugfix: Pressing the "Delete" key during a datagrid cell edit action caused deletion of the entire row.
- Bugfix in sale window for the exception which is fired if the users cancels before adding any items to a sale.
- CenterOwner bugfix for user controls (Parent of the user control is not directly a window).
- Mouse double click of the products datagrid opens details of the product now.
- Stock control after expiration date modifications
- Bugfix of the sales user control: Expiration Date combobox selection changes weren't updating the related stocks of the expiration dates.

+++ IremEczOtomasyonu/SaleListWindow.xaml.cs
@@ -78,7 +78,7 @@ namespace IremEczOtomasyonu
 
         private void ProductSalesDataGrid_PreviewKeyDown(object sender, KeyEventArgs e)
         {
-            if (e.Key == Key.Delete)
+            if (e.OriginalSource is DataGridCell && e.Key == Key.Delete)
             {
                 DeleteSelectedProductSale();
                 e.Handled = true;
